Adapt variable neutralisation
